Update scripts to ask for root access

+++ rebuild_services.sh
@@ -1,4 +1,15 @@
 #!/usr/bin/bash
+$user=$(whoami)
+if [ $user != "root" ]; then
+    echo "You must run this script as root"
+    exit 1
+fi
+
 git pull
 docker compose down
+
+sudo rm -rf ./automatization/build/*
+sudo rm -rf ./backend/build/*
+sudo rm -rf ./frontend/build/*
+
 docker compose up -d --build
