+++ update_services.sh
@@ -1,4 +1,10 @@
 #!/usr/bin/bash
+$user=$(whoami)
+if [ $user != "root" ]; then
+    echo "You must run this script as root"
+    exit 1
+fi
+
 git pull
 
 sudo rm -rf ./automatization/build/*
